+++ chat-widget/src/components/startchaterrorpanestateful/interfaces/IStartChatErrorPaneControlProps.ts
@@ -1,0 +1,10 @@
+export interface IStartChatErrorPaneControlProps {
+    id?: string;
+    role?: string;
+    dir?: "rtl" | "ltr" | "auto";
+    hideIcon?: boolean;
+    hideTitle?: boolean;
+    hideSubtitle?: boolean;
+    titleText?: string;
+    subtitleText?: string;
+}